Alternative Interface First Commit
I made a dropdown menu based interface with some example choice categories. These may not end up being the final choices depending on what is decided. Since the form is non-functional, I disabled the submit button.

Also added a skeleton API endpoint, but the actual querying still needs to be written as I wasn't exactly sure of the formatting for that with supabase.